Factor out duplicated code
