+++ src/components/modules/epics/ContributionsEpicReminder.tsx
@@ -1,26 +1,10 @@
 // --- Imports --- //
 
 import React, { useState } from 'react';
-import { addCookie } from '../../../lib/cookies';
 import { OneOffSignupRequest } from '../../../api/supportRemindersApi';
 import { ContributionsEpicReminderSignedIn } from './ContributionsEpicReminderSignedIn';
 import { ContributionsEpicReminderSignedOut } from './ContributionsEpicReminderSignedOut';
-
-// --- Utils --- //
-
-const dateDiff = (start: Date, end: Date): number => {
-    const twentyFourHours = 86400000;
-    return Math.round((end.valueOf() - start.valueOf()) / twentyFourHours);
-};
-
-const addContributionReminderCookie = (reminderDateString: string): void => {
-    const today = new Date();
-    const reminderDate = new Date(Date.parse(reminderDateString));
-
-    addCookie('gu_epic_contribution_reminder', '1', dateDiff(today, reminderDate));
-};
-
-const createOneOffReminderEndpoint = 'https://support.theguardian.com/reminders/create/one-off';
+import { addContributionReminderCookie } from './utils/reminders';
 
 // --- Types --- //
 
@@ -38,9 +22,15 @@ export enum ReminderStatus {
     Completed = 'Completed',
 }
 
+// --- Consts --- //
+
+const CREATE_ONE_OFF_REMINDER_ENDPOINT = 'https://support.theguardian.com/reminders/create/one-off';
+
 const REMINDER_PLATFORM = 'WEB';
 const REMINDER_COMPONENT = 'EPIC';
 const REMINDER_STAGE = 'PRE';
+
+// --- Components --- //
 
 export const ContributionsEpicReminder: React.FC<ContributionsEpicReminderProps> = ({
     initialEmailAddress,
@@ -60,7 +50,7 @@ export const ContributionsEpicReminder: React.FC<ContributionsEpicReminderProps>
         };
 
         setReminderStatus(ReminderStatus.Submitting);
-        fetch(createOneOffReminderEndpoint, {
+        fetch(CREATE_ONE_OFF_REMINDER_ENDPOINT, {
             body: JSON.stringify(reminderSignupData),
             method: 'POST',
             headers: {
